Facial featrue multi task classification/Resnet-50/requirements.txt
torch
torchvision
pandas
scikit-learn
Pillow
tqdm
